portfolio/src/main/java/com/google/sps/servlets/DeleteDataServlet.java
package com.google.sps.servlets;

import java.io.IOException;
import javax.servlet.annotation.WebServlet;
import javax.servlet.http.HttpServlet;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;

@WebServlet("/delete-data")
public class DeleteDataServlet extends HttpServlet {
  @Override
  public void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
    CommentsHandler commentsHandler = new CommentsHandler();
    // get the key of comment that has to be deleted
    String commentKey = request.getParameter("commentKey");
    commentsHandler.deleteComment(commentKey);
    response.sendRedirect("./index.html");
  }
}
